Add files via upload
Hi, here is my Python Assignment of Linked List. I hope that this will help for those who are facing issues on understanding the Linked List Concept. We have to done Following Tasks
Exercises
1. Write a python program to create a linked list of numbers in it. Your linked list should 
contain following functions.
a. addatStart(): to add a number at the start of the linked list.
b. removeFromStart(): to delete a number from the start of the list.
c. displayNumbers(): to print all the numbers present in the linked list.
2. Write a python program to create a linked list to store information of different persons. 
Attributes of the class person include name, age and city. Your linked list should contain 
following functions.
a. addatEnd(): to add a person at the end of the linked list.
b. removeFromEnd(): to remove a person value form the end of the linked list.
c. displayPersons(): to print all the persons present in the linked list.
3. Write a python program to create a linked list of books. Attributes of the class book 
include title, author name and edition. Your linked list should contain following functions:
a. addatStart(): to add a book at the start of the linked list.
b. removeFromStart(): to delete a book from the start of the list.
c. addatEnd(): to add a book at the end of the linked list.
d. removeFromEnd(): to remove a book form the end of the linked list.
e. displayBooks(): to print all the books present in the linked list.
f. searchBook(): search book in the linked list based on the title. It should 
return true if book with the given title is present in the list otherwise return 
false.
g. booksCount(): return number of books present in the linked list.
h. Exit(): stop the program.
i. Provide a menu to use all the functions of the linked list.
4. Write a python program to create a linked list of subjects. Attributes of the subject class 
include title, credit hours, semester and instructor. Linked list should contain following 
functions:
a. addatStart(): to add a subject at the start of the linked list.
b. removeFromStart(): to delete a subject from the start of the list.
c. addatEnd(): to add a subject at the end of the linked list.
d. removeFromEnd(): to remove a subject form the end of the linked list.
e. addatPosition(): to add a subject at the given position.
f. removeFromPosition(): to delete a subject from the given position.
g. valueatIPosition(): to display the subject present at the given Position.
h. searchSubject(): search subject in the linked list based on the title. It 
should return true if subject with the given title is present in the list otherwise 
return false. 
i. Size(): to print number of subjects present in the linked list.
j. PrintSubjects(): print attributes of all the subject present in the linked list.
k. Exit(): stop the program
5. Provide a menu to use the linked list developed in Q5. Take input from a-l and then call 
appropriate function. For example, if user enters 
‘c’, then take attributes of subject as input, create object using the attributes and then 
add this object at the end of the linked list.
‘f’, then take position as input and remove the subject present at the given position. 
Note: for the parts e, f and g, take position as input. Its value ranges from 1-size of the linked list